fix: cmd/agent Выделить получение статистик в пакет internal/stats
