+++ cmd/agent/agent.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/thefrol/kysh-kysh-meow/internal/scheduler"
+	"github.com/thefrol/kysh-kysh-meow/internal/stats"
 	"github.com/thefrol/kysh-kysh-meow/internal/storage"
 )
 
@@ -23,8 +24,8 @@ func main() {
 	//собираем данные раз в pollingInterval
 	c.AddJob(time.Duration(*pollIntervalSeconds)*time.Second, func() {
 		//Обновляем данные в хранилище
-		fetchMemStats(store)
-		fetchAdditionalStats(store)
+		stats.FetchMemStats(store)
+		stats.FetchAdditionalStats(store)
 		// Увеличиваем PollCount
 		incrementCounter(store, metricPollCount)
 	})
