+++ cmd/agent/pollcount.go
@@ -5,6 +5,10 @@ import (
 	"github.com/thefrol/kysh-kysh-meow/internal/storage"
 )
 
+const (
+	metricPollCount = "PollCount"
+)
+
 // incrementCounter обновляет PollCount счетчик в хранилище, добавляет ему единицу
 func incrementCounter(store storage.Storager, name string) {
 	count, _ := store.Counter(metricPollCount)
