+++ internal/stats/fetch_test.go
@@ -1,11 +1,10 @@
 // fetchStats собирает метрики из памяти, и выдает их удобной мапой
-package main
+package stats
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/thefrol/kysh-kysh-meow/internal/metrica"
 	"github.com/thefrol/kysh-kysh-meow/internal/storage"
 )
 
@@ -61,7 +60,7 @@ func Test_fetchMemStats(t *testing.T) {
 	for _, tt := range tests {
 
 		t.Run(tt.name, func(t *testing.T) {
-			fetchMemStats(tt.args.store)
+			FetchMemStats(tt.args.store)
 			if tt.memValuesCount >= 0 {
 				assert.Equal(t, tt.memValuesCount, CountValues(tt.args.store))
 			}
@@ -89,76 +88,17 @@ func Test_fetchAdditionalStats(t *testing.T) {
 			name:        "all metrics in place",
 			args:        args{store: storage.New()},
 			wantErr:     false,
-			fieldsFound: []string{metricRandomValue},
+			fieldsFound: []string{randomValueName},
 		},
 	}
 	for _, tt := range tests {
 
 		t.Run(tt.name, func(t *testing.T) {
-			fetchAdditionalStats(tt.args.store)
+			FetchAdditionalStats(tt.args.store)
 			for _, v := range tt.fieldsFound {
 				_, gaugeFound := tt.args.store.Gauge(v)
 				_, counterFound := tt.args.store.Counter(v)
 				assert.Truef(t, gaugeFound || counterFound, "Not found metric %v", v)
-			}
-
-		})
-	}
-}
-
-func Test_updateCounter(t *testing.T) {
-	type args struct {
-		store       storage.Storager
-		counterName string
-	}
-	tests := []struct {
-		name          string
-		args          args
-		wantErr       bool
-		runsCount     int
-		counterValues map[string]metrica.Counter //проверяет значение этого счетчика
-		foundValues   map[string]bool            // проверяет наличие такого счетчика,
-	}{
-		{
-			name:          "zero runs",
-			args:          args{store: storage.New(), counterName: metricPollCount},
-			wantErr:       false,
-			runsCount:     0,
-			counterValues: map[string]metrica.Counter{},
-			foundValues:   map[string]bool{metricPollCount: false},
-		},
-		{
-			name:          "one run",
-			args:          args{store: storage.New(), counterName: metricPollCount},
-			wantErr:       false,
-			runsCount:     1,
-			counterValues: map[string]metrica.Counter{metricPollCount: metrica.Counter(1)},
-			foundValues:   nil,
-		},
-		{
-			name:          "three runs",
-			args:          args{store: storage.New(), counterName: metricPollCount},
-			wantErr:       false,
-			runsCount:     3,
-			counterValues: map[string]metrica.Counter{metricPollCount: metrica.Counter(3)},
-			foundValues:   nil,
-		},
-	}
-	for _, tt := range tests {
-
-		t.Run(tt.name, func(t *testing.T) {
-
-			for i := 0; i < tt.runsCount; i++ {
-				incrementCounter(tt.args.store, tt.args.counterName)
-			}
-			for k, expect := range tt.counterValues {
-				real, ok := tt.args.store.Counter(k)
-				assert.Truef(t, ok, "Not found counter %v", k)
-				assert.Equal(t, expect, real)
-			}
-			for k, expect := range tt.foundValues {
-				_, ok := tt.args.store.Counter(k)
-				assert.Equalf(t, expect, ok, "Counter %v should be found(%v), but got %v", k, expect, ok)
 			}
 
 		})
